% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kingman_mean.R
\name{bal_Fmat}
\alias{bal_Fmat}
\title{Compute the 'most balanced' F-matrix}
\usage{
bal_Fmat(n = NULL)
}
\arguments{
\item{n}{The number of leaves}
}
\description{
Compute the F-fmatrix of the most balanced tree
}
