% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_Fmats.R
\name{distance_Fmat}
\alias{distance_Fmat}
\title{Evaluate Distance between Fmatrices}
\usage{
distance_Fmat(Fmat1, Fmat2, dist = c("l2", "l1"), scale = NULL)
}
\arguments{
\item{dist}{}
}
\description{
Evaluate Distance between Fmatrices
}
