% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_Fmats.R
\name{distance_Fmat_weighted}
\alias{distance_Fmat_weighted}
\title{Evaluate weighted Distance between Fmatrices}
\usage{
distance_Fmat_weighted(wFmat1, wFmat2, dist = c("l2", "l1"))
}
\arguments{
\item{dist}{}
}
\description{
Evaluate weighted Distance between Fmatrices
}
