% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frechet_var.R
\name{frechet_var_sample}
\alias{frechet_var_sample}
\title{frechet variance of sample}
\usage{
frechet_var_sample(F.list, dist = c("l2", "l1"))
}
\arguments{
\item{dist}{}
}
\description{
frechet variance of sample
}
