% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minor_wrappers.R
\name{heatF}
\alias{heatF}
\title{Plot Heatmap of F-matrix}
\usage{
heatF(m, ...)
}
\arguments{
\item{m}{F-matrix}

\item{...}{Arguments to be passed to heatmap}
}
\description{
Plot heatmap of F-matrix
}
