% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_encod_hetero.R
\name{is.hEncod}
\alias{is.hEncod}
\title{Check if a string is a valid hEncod}
\usage{
is.hEncod(hEncod, digits.tol = 6)
}
\arguments{
\item{hEncod}{}
}
\description{
Check if a string correctly encodes a heterochronous ranked tree shape
}
