% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rEncod_maliet.R
\name{rEncod.single.Maliet2018}
\alias{rEncod.single.Maliet2018}
\title{Generate a random encoding}
\usage{
rEncod.single.Maliet2018(
  n,
  b = 0,
  alpha = 1,
  eta = 1,
  type = c("encod", "hEncod")
)
}
\arguments{
\item{n}{n+1 is the number of tips of the tree}

\item{b}{beta parameter in beta model of Maliet2018.
Default b = 0 gives the Yule / Kingman model}

\item{alpha}{alpha parameter in Maliet2018, age-richness index
Default is alpha = 1.}

\item{eta}{eta parameter in Maliet2018, abundance-richness index (Not implemented!)
Default would be eta = 1.}

\item{type}{flag to switch between isochronous and hetero
chronous versions of encodings}
}
\description{
Generate a random encoding according to Maliet-Aldous beta splitting model
}
