% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rEncod.R
\name{rEncod.single.SV2016}
\alias{rEncod.single.SV2016}
\title{Generate a random encoding}
\usage{
rEncod.single.SV2016(n, b = 0, a = b, type = c("encod", "hEncod"))
}
\arguments{
\item{n}{n+1 is the number of tips of the tree}

\item{b}{beta parameter in BF model of SV2016.
Default b = 0 gives the Yule / Kingman model}

\item{a}{alpha parameter in BF model of SV 2016.
Default is a = b for BF model.}

\item{type}{flag to switch between isochronous and hetero
chronous versions of encodings}
}
\description{
Generate a random encoding according to Blum-Francoid beta splitting model
}
