% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rEncod_uniform.R
\name{rEncod.single.uniform}
\alias{rEncod.single.uniform}
\title{Generate a random encoding}
\usage{
rEncod.single.uniform(n, type = c("encod", "hEncod"))
}
\arguments{
\item{n}{n+1 is the number of tips of the tree}

\item{type}{flag to switch between isochronous and hetero
chronous versions of encodings}
}
\description{
Generate a random encoding uniformly
}
