% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frechet_mean.R
\name{sa_mean}
\alias{sa_mean}
\title{Estimate the Frechet Mean of a list of F-matrices using Simulated Annealing}
\usage{
sa_mean(
  F.list,
  weights = NULL,
  temp.schedule = c("exp", "lin", "log"),
  chain = c("encod", "matching"),
  track = FALSE,
  init.temp = NULL,
  alpha = NULL,
  max.iter = NULL
)
}
\arguments{
\item{F.list}{list of F-matrices}

\item{weights}{weights for F.list}

\item{temp.schedule}{temperature schedule, one of "exp", "linear", or "log"}

\item{track}{boolean, returns tracked stuff if TRUE}
}
\description{
This function estimates the Frechet Mean of a given list of F-matrices using Simulated Annealing
}
