% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minor_wrappers.R
\name{scaling_vector}
\alias{scaling_vector}
\title{Create a scaling vector for distances}
\usage{
scaling_vector(n, type = 0)
}
\arguments{
\item{type}{}
}
\description{
Create a scaling vector for distances
}
