% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary_order.R
\name{signed_dist}
\alias{signed_dist}
\title{Evaluate signed distance of F-matrix}
\usage{
signed_dist(
  Fmat,
  d = c("l2", "l1"),
  precomp = precomp_bal_unb_kin(ncol(Fmat) + 1)
)
}
\arguments{
\item{precomp}{Default is NULL.
Else, list of Fbal, Funb, Fkin
which are precomputed}
}
\description{
Evaluate signed distance of F-matrix
}
