% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coal_times_fromto_wt.R
\name{times_from_wt}
\alias{times_from_wt}
\title{Get coalescent times from weight matrix}
\usage{
times_from_wt(wt, type = c("1", "2"))
}
\arguments{
\item{wt}{}
}
\description{
Get coalescent times from matrix of weights
corresponding to F-matrix
}
