% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imported-tree_from_F.R
\name{tree_from_F}
\alias{tree_from_F}
\title{Generate tree in phylo form from F-matrix}
\usage{
tree_from_F(matF, coal_times)
}
\arguments{
\item{coal_times}{}
}
\description{
This functions generates a tree in phylo format to plot from an F-matrix
}
