% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kingman_mean.R
\name{unb_Fmat}
\alias{unb_Fmat}
\title{Compute the 'most unbalanced' F-matrix}
\usage{
unb_Fmat(n = NULL)
}
\arguments{
\item{n}{The number of leaves}
}
\description{
Compute the F-fmatrix of the most unbalanced tree
}
