% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coal_times_fromto_wt.R
\name{wt_from_times}
\alias{wt_from_times}
\title{Get weight matrix from coalescent times}
\usage{
wt_from_times(coal_times, type = c("2", "1"))
}
\arguments{
\item{type}{}
}
\description{
Get matrix of weights corresponding to F-matrix for a given
vector of coalescent times
}
