% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frechet_var.R
\name{frechet_var_pop}
\alias{frechet_var_pop}
\title{frechet variance of population}
\usage{
frechet_var_pop(F.list, probs, dist = c("l1", "l2"))
}
\arguments{
\item{dist}{}
}
\description{
frechet variance of population
}
